+++ lib/system_test/include/metashell/system_test/code_completion_result.hpp
@@ -56,6 +56,7 @@ namespace metashell
     json_string to_json_string(const code_completion_result& r_);
 
     bool operator==(const code_completion_result& r_, const json_string& s_);
+    bool operator!=(const code_completion_result& r_, const json_string& s_);
   }
 }